// src/generated/deployments.ts
export const deploymentManifest = {
  "$schema": "./blockchains.schema.json",
  "schema_version": 1,
  "active_network": "mainnet",
  "deployer_address": "0xD07B1A5110f0042d9d38E66FFc1E9Ee07F1C9025",
  "robinhood": {
    "mainnet": {
// ...
      "contracts": {
// ...
        "implementations": {
          "launchpad": null,
          "fee_controller": {
            "address": "0x598fd4da33e893f4a2b195b38a9838a27dbd7352",
            "runtime_code_hash": "0x6303dde8ed462b9b1c078dc75810f10115d58b991145f4153344e7c6242e855f"
          }
        },
        "escrow_implementation": "0xb0743a0c4740837245cd6b6b96c27d1379571b51"
      }
    },
// ...
  }
} as const;